Use an abstract predicate function in ForkByKeyProvider
