+++ provider/datagen/src/registry.rs
@@ -114,7 +114,7 @@ macro_rules! create_datagen_provider {
     ($source_data:expr, [ $($constructor:path),+, ]) => {{
         let __source = &$source_data;
         icu_provider_adapters::make_forking_provider!(
-            icu_provider_adapters::fork::by_key::ForkByKeyProvider,
+            icu_provider_adapters::fork::ForkByKeyProvider::new,
             [
                 icu_provider::hello_world::HelloWorldProvider,
                 $(<$constructor>::from(__source)),+,
@@ -188,7 +188,7 @@ macro_rules! create_datagen_provider {
     ($source_data:expr, [ $($constructor:path),+, ]) => {{
         let __source = &$source_data;
         icu_provider_adapters::make_forking_provider!(
-            icu_provider_adapters::fork::by_key::ForkByKeyProvider,
+            icu_provider_adapters::fork::ForkByKeyProvider::new,
             [
                 icu_provider::hello_world::HelloWorldProvider,
                 $(<$constructor>::from(__source)),+,
